% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_client.R
\name{adbc_connect}
\alias{adbc_connect}
\title{Get a connection to flight sql server}
\usage{
adbc_connect(cs)
}
\arguments{
\item{cs}{connectionstring, by default duckhttp_adbc_con_default}
}
\value{
an adbc connection object
}
\description{
Get a connection to flight sql server
}
