% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_client.R
\name{adbc_disconnect}
\alias{adbc_disconnect}
\title{Disconnect from the Flight SQL server}
\usage{
adbc_disconnect(adbc_con)
}
\arguments{
\item{adbc_con}{an adbc connection object}
}
\description{
Disconnect from the Flight SQL server
}
