% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_client.R
\name{adbc_query}
\alias{adbc_query}
\title{Query a Flight SQL server}
\usage{
adbc_query(adbc_con, query, silent = FALSE)
}
\arguments{
\item{adbc_con}{an adbc connection object}

\item{query}{the sql query}

\item{silent}{boolean to indicate if progress should be reported}
}
\description{
Query a Flight SQL server
}
