% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_client.R
\name{adbc_table}
\alias{adbc_table}
\title{Read full table from remote Flight SQL server}
\usage{
adbc_table(adbc_con, tablename)
}
\arguments{
\item{adbc_con}{an object returned from adbc_connect}

\item{tablename}{the name of a table to return data from}
}
\value{
a tibble
}
\description{
Read full table from remote Flight SQL server
}
