% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_client.R
\name{adbc_tables}
\alias{adbc_tables}
\title{Enumerate tables at remote Flight SQL server}
\usage{
adbc_tables(adbc_con)
}
\arguments{
\item{adbc_con}{an object returned from adbc_connect}
}
\value{
a character vector of available tables
}
\description{
Enumerate tables at remote Flight SQL server
}
