% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_client.R
\name{duckhttp_con}
\alias{duckhttp_con}
\title{Create a connection to a duckdb httpserver instance}
\usage{
duckhttp_con(connectionstring = NULL, quiet = TRUE, api_key)
}
\arguments{
\item{connectionstring}{the connectionstring to use for the duckdb httpserver}

\item{quiet}{a boolean indicating whether some messages should be displayed}

\item{api_key}{a character only provided if using an API key}
}
\value{
a connection object which can be used with the other functions
}
\description{
Create a connection to a duckdb httpserver instance
}
\details{
The connectionstring should follow this pattern:
http(s)://user:pass@hostname/path/

Examples:

http://localhost/duckserve/
https://user:pass@some.server.tld/duckserve/

When not using basic auth (like above), use the api_key parameter
}
