% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_client.R
\name{duckhttp_ls}
\alias{duckhttp_ls}
\title{List available tables}
\usage{
duckhttp_ls(con = duckhttp_con(quiet = TRUE))
}
\arguments{
\item{con}{the connection for the duckdb httpserver, see duckhttp_con()}
}
\value{
a character vector of table names
}
\description{
List available tables
}
