% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_client.R
\name{duckhttp_ping}
\alias{duckhttp_ping}
\title{Check that a duckdb httpserver instance is available}
\usage{
duckhttp_ping(con = duckhttp_con(quiet = TRUE))
}
\arguments{
\item{con}{the connection for the duckdb httpserver, see duckhttp_con()}
}
\value{
TRUE if available, otherwise FALSE
}
\description{
Check that a duckdb httpserver instance is available
}
