% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_client.R
\name{duckhttp_read}
\alias{duckhttp_read}
\title{Read data from duckdb running httpserver extension}
\usage{
duckhttp_read(con = duckhttp_con(quiet = TRUE), query = NULL)
}
\arguments{
\item{con}{the connection, by default from duckhttp_con()}

\item{query}{the sql query to issue against the duckdb httpserver}
}
\value{
a tibble with results from the query
}
\description{
Read data from duckdb running httpserver extension
}
