% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_client.R
\name{duckhttp_read_csv}
\alias{duckhttp_read_csv}
\title{Issue a query against a remote duckdb instance running the httpserver extension (with CSV formatted results converted to data frame)}
\usage{
duckhttp_read_csv(con = duckhttp_con(quiet = TRUE), query = NULL)
}
\arguments{
\item{con}{the connection, by default from duckhttp_con()}

\item{query}{the sql query to issue against the duckdb httpserver}
}
\value{
a tibble with results from the query
}
\description{
Issue a query against a remote duckdb instance running the httpserver extension (with CSV formatted results converted to data frame)
}
