% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_client.R
\name{duckhttp_ui}
\alias{duckhttp_ui}
\title{Open the duckdb httpserver web UI}
\usage{
duckhttp_ui(con = duckhttp_con(quiet = FALSE))
}
\arguments{
\item{con}{the connection object from duckhttp_con()}
}
\value{
the url base for the web ui
}
\description{
Open the duckdb httpserver web UI
}
